requirements.txt

Flask==2.0.1
snowflake-connector-python==2.7.9
